[FIX] Fix float parsing error in config files

+++ cs336_basics/core/loss.py
@@ -25,9 +25,7 @@ def cross_entropy_loss(
     logits = logits - max_logits
 
     log_sum = torch.log(torch.sum(torch.exp(logits), dim=-1, keepdim=True))
-    prob_at_true_label = torch.gather(
-        logits, dim=-1, index=targets.unsqueeze(-1)
-    )  # (B, ..., L, 1)
+    prob_at_true_label = torch.gather(logits, dim=-1, index=targets.unsqueeze(-1).long())  # (B, ..., L, 1)
     loss_in_batch = log_sum - prob_at_true_label
 
     match reduction:
